Initialize ChatOpenAI with str rather than SecretStr
For some reason the cloud instance was throwing an exception complaining
about the api_key parameter not being a str (it was a SecretStr).
Interestingly, running locally, in my dev environment was working fine
with the SecretStr. But the error was quite explicit in the Cloud Run
logs. So, I converted the token to a regular str and everything is
working fine now.